Implementation of imap_quote() with fixed input buffer
Quotes src string and copies to dst buffer, which must be at least twice
the length of the src string to account for the worst case password
consisting of only \ and " chars, plus three for the surrounding quotes
and the terminal null byte.

This patch only quotes input for the LOGIN imap_exec(), although there
exist other imap_exec() calls with \"%s\" in their format strings.