make paths relative to CWD, after all
the test suite actually relies on it. it would be possible to adjust it,
but there is not much reason to make paths relative to HOME (as we
support convenient tilde expansion). so use the least invasive approach,
which is simply the old behavior. adjust the documentation accordingly.

This reverts commit d53205fde766ec7abf69088ee247e316d9a0118a.
(cherry picked from commit b802c7d4ed6eb09bd6cba94315d43158e20fdec6)